+++ src/Akeneo/Channel/Bundle/Controller/InternalApi/CurrencyController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\EnrichBundle\Controller\Rest;
+namespace Akeneo\Channel\Bundle\Controller\InternalApi;
 
 use Akeneo\Channel\Component\Repository\CurrencyRepositoryInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
