+++ src/Akeneo/Channel/Bundle/Controller/UI/CurrencyController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\EnrichBundle\Controller;
+namespace Akeneo\Channel\Bundle\Controller\UI;
 
 use Akeneo\Channel\Component\Exception\LinkedChannelException;
 use Akeneo\Channel\Component\Model\Currency;
