+++ src/Akeneo/Channel/Bundle/Resources/config/routing.yml
@@ -17,3 +17,11 @@ external_api_channel:
 external_api_currency:
     resource: '@AkeneoChannelBundle/Resources/config/routing/external_api/currency.yml'
     prefix: /api/rest/v1
+
+currency_internal_api:
+    resource: '@AkeneoChannelBundle/Resources/config/routing/internal_api/currency.yml'
+    prefix: /configuration/currency
+
+currency_ui:
+    resource: '@AkeneoChannelBundle/Resources/config/routing/ui/currency.yml'
+    prefix: /configuration/currency
