+++ src/Akeneo/Channel/Bundle/Resources/config/routing/internal_api/currency.yml
@@ -1,0 +1,4 @@
+pim_enrich_currency_rest_index:
+    path: /rest
+    defaults: { _controller: pim_enrich.controller.rest.currency:indexAction }
+    methods: [GET]
