+++ src/Akeneo/Channel/Bundle/Resources/config/routing/ui/currency.yml
@@ -6,8 +6,3 @@ pim_enrich_currency_toggle:
     defaults: { _controller: pim_enrich.controller.currency:toggleAction }
     requirements:
         id: \d+
-
-pim_enrich_currency_rest_index:
-    path: /rest
-    defaults: { _controller: pim_enrich.controller.rest.currency:indexAction }
-    methods: [GET]
