+++ src/Akeneo/Pim/Enrichment/Bundle/Controller/InternalApi/ProductCommentController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\EnrichBundle\Controller\Rest;
+namespace Akeneo\Pim\Enrichment\Bundle\Controller\InternalApi;
 
 use Akeneo\Pim\Enrichment\Bundle\Form\Type\CommentType;
 use Akeneo\Pim\Enrichment\Component\Comment\Builder\CommentBuilder;
