+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/routing.yml
@@ -16,3 +16,35 @@ external_api_category:
 external_api_product:
     resource: '@AkeneoPimEnrichmentBundle/Resources/config/routing/external_api/product.yml'
     prefix: /api/rest/v1
+
+category_internal_api:
+    resource: '@AkeneoPimEnrichmentBundle/Resources/config/routing/internal_api/category.yml'
+    prefix: /enrich/category
+
+product_internal_api:
+    resource: '@AkeneoPimEnrichmentBundle/Resources/config/routing/internal_api/product.yml'
+    prefix: /enrich/product
+
+group_internal_api:
+    resource: '@AkeneoPimEnrichmentBundle/Resources/config/routing/internal_api/group.yml'
+    prefix: /enrich/group
+
+sequential_edit_internal_api:
+    resource: '@PimEnrichBundle/Resources/config/routing/sequential_edit/rest.yml'
+    prefix: /rest/sequential_edit
+
+value_internal_api:
+    resource: '@AkeneoPimEnrichmentBundle/Resources/config/routing/internal_api/value.yml'
+    prefix: /rest/value
+
+mass_edit_rest:
+    resource: '@AkeneoPimEnrichmentBundle/Resources/config/routing/internal_api/mass_edit.yml'
+    prefix: /rest/mass_edit
+
+media:
+    resource: '@PimEnrichBundle/Resources/config/routing/media.yml'
+    prefix: /media
+
+categorytree_ui:
+    resource: '@AkeneoPimEnrichmentBundle/Resources/config/routing/ui/categorytree.yml'
+    prefix: /enrich/product-category-tree
