+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/routing/internal_api/mass_edit.yml
