+++ src/Akeneo/Platform/Bundle/ImportExportBundle/Resources/config/routing.yml
@@ -13,3 +13,11 @@ import_report:
 export_report:
     resource: '@PimImportExportBundle/Resources/config/routing/export_execution.yml'
     prefix: /spread/export_execution
+
+job_instance_internal_api:
+    resource: '@PimEnrichBundle/Resources/config/routing/job_instance.yml'
+    prefix: /job-instance
+
+job_execution_rest:
+    resource: '@PimEnrichBundle/Resources/config/routing/job_execution.yml'
+    prefix: /job-execution
