+++ src/Akeneo/Platform/Bundle/UIBundle/Resources/config/controllers.yml
@@ -7,3 +7,8 @@ services:
         arguments:
             - '@doctrine'
             - '@pim_reference_data.registry'
+
+    pim_enrich.controller.rest.form_extension:
+        class: 'Akeneo\Platform\Bundle\UIBundle\Controller\InternalApi\FormExtensionController'
+        arguments:
+            - '@pim_enrich.provider.form_extension'
