+++ src/Akeneo/Platform/Bundle/UIBundle/Resources/config/routing/internal_api/form_extensions.yml
