+++ src/Pim/Bundle/EnrichBundle/Resources/config/routing/product.yml
@@ -21,77 +21,3 @@ pim_enrich_product_listcategories:
         id: '[0-9a-f]+'
         categoryId: \d+
 
-pim_enrich_product_rest_get:
-    path: /rest/{id}
-    defaults: { _controller: pim_enrich.controller.rest.product:getAction }
-    requirements:
-        id: '[0-9a-f]+'
-    methods: [GET]
-
-pim_enrich_product_rest_index:
-    path: /rest/
-    defaults: { _controller: pim_enrich.controller.rest.product:indexAction }
-    methods: [GET]
-
-pim_enrich_product_rest_create:
-    path: /rest
-    defaults: { _controller: pim_enrich.controller.rest.product:createAction }
-    methods: [POST]
-
-pim_enrich_product_rest_post:
-    path: /rest/{id}
-    defaults: { _controller: pim_enrich.controller.rest.product:postAction }
-    requirements:
-        id: '[0-9a-f]+'
-    methods: [POST]
-
-pim_enrich_product_rest_remove:
-    path: /rest/{id}
-    defaults: { _controller: pim_enrich.controller.rest.product:removeAction, _format: json }
-    requirements:
-        id: '[0-9a-f]+'
-    methods: [DELETE]
-
-pim_enrich_product_category_rest_list:
-    path: /rest/{id}/categories
-    defaults: { _controller: pim_enrich.controller.rest.product_category:listAction }
-    requirements:
-        id: '[0-9a-f]+'
-    methods: [GET]
-
-pim_enrich_product_history_rest_get:
-    path: /rest/product/{entityId}/history
-    defaults: { _controller: pim_enrich.controller.rest.versioning:getAction, entityType: product }
-    requirements:
-        entityId: '[0-9a-f]+'
-    methods: [GET]
-
-pim_enrich_product_remove_attribute_rest:
-    path: /rest/{id}/{attributeId}
-    defaults: { _controller: pim_enrich.controller.rest.product:removeAttributeAction }
-    requirements:
-        id: '[0-9a-f]+'
-        attributeId: '\d+'
-    methods: [DELETE]
-
-pim_enrich_product_comments_rest_get:
-    path: /{id}/comments
-    defaults: { _controller: pim_enrich.controller.rest.product_comment:getAction }
-    requirements:
-        id: '[0-9a-f]+'
-    methods: [GET]
-
-pim_enrich_product_comments_rest_post:
-    path: /{id}/comments
-    defaults: { _controller: pim_enrich.controller.rest.product_comment:postAction }
-    requirements:
-        id: '[0-9a-f]+'
-    methods: [POST]
-
-pim_enrich_product_comment_reply_rest_post:
-    path: /{id}/comments/{commentId}
-    defaults: { _controller: pim_enrich.controller.rest.product_comment:postReplyAction }
-    requirements:
-        id: '[0-9a-f]+'
-        commentId: '[0-9a-f]+'
-    methods: [POST]
